femnist/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
blessed==1.21.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorful==0.5.7
contourpy==1.3.2
cryptography==41.0.7
cycler==0.12.1
Cython==3.1.2
datasets==3.1.0
dill==0.3.8
distlib==0.4.0
filelock==3.18.0
flatbuffers==25.2.10
flwr==1.5.0
flwr-datasets==0.4.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.9.0
gast==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gpustat==1.1.1
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.7
huggingface-hub==0.34.3
hydra-core==1.3.2
idna==3.10
iterators==0.0.2
jax==0.4.38
jaxlib==0.4.38
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==2.12.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
ml_dtypes==0.5.3
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
numpy==1.24.3
nvidia-ml-py==13.580.65
oauthlib==3.3.1
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==3.11.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
py-spy==0.4.1
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==1.10.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ray==2.6.3
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smart_open==7.3.0.post1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
xxhash==3.5.0
yarl==1.20.1
